+++ utils/ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: hyeonjan <hyeonjan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/10 20:06:44 by yenawee           #+#    #+#             */
-/*   Updated: 2022/07/12 21:18:01 by hyeonjan         ###   ########.fr       */
+/*   Updated: 2022/07/12 22:39:07 by hyeonjan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
