+++ utils/ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: hyeonjan <hyeonjan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/23 17:29:18 by hyeonjan          #+#    #+#             */
-/*   Updated: 2022/07/12 20:28:14 by hyeonjan         ###   ########.fr       */
+/*   Updated: 2022/07/12 22:38:51 by hyeonjan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
